Reorganize constants into global constants file, move subsystem specific constants into subsystem class
