+++ src/main/java/frc/robot/commands/AutoAimCommand.java
@@ -1,11 +1,9 @@
  package frc.robot.commands;
 
-import com.ctre.phoenix6.mechanisms.swerve.SwerveRequest;
-import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.FieldConstants;
-import frc.robot.constants.VisionConstants;
+import frc.robot.constants.ConfiguratorConstants;
+import frc.robot.subsystems.VisionSubsystem;
 import frc.robot.subsystems.drivetrain.CommandSwerveDrivetrain;
 import frc.robot.util.AllianceFlipUtil;
 import frc.robot.util.LimelightHelpers;
@@ -40,11 +38,11 @@ public class AutoAimCommand extends Command {
     @Override
     public void initialize() {
 
-        currentTag = LimelightHelpers.getFiducialID(VisionConstants.LIMELIGHT_NAME);
+        currentTag = LimelightHelpers.getFiducialID(VisionSubsystem.LIMELIGHT_NAME);
 
         poseY = drivetrain.getState().Pose.getX();
         Translation2d speakerCenter = AllianceFlipUtil.apply(
-                FieldConstants.Speaker.centerSpeakerOpening.toTranslation2d()
+                ConfiguratorConstants.Speaker.centerSpeakerOpening.toTranslation2d()
         );
 
         poseAimRequest.setPointToFace(speakerCenter);
@@ -52,7 +50,7 @@ public class AutoAimCommand extends Command {
 
     @Override
     public void execute() {
-        if(LimelightHelpers.getFiducialID(VisionConstants.LIMELIGHT_NAME) == currentTag) {
+        if(LimelightHelpers.getFiducialID(VisionSubsystem.LIMELIGHT_NAME) == currentTag) {
             drivetrain.setControl(
                     poseAimRequest.withVelocityX(xVelSupplier.getAsDouble() * 1.5).withVelocityY(yVelSupplier.getAsDouble() * 1.5)
             );
