+++ src/main/java/frc/robot/commands/HandoffCommandGroup.java
@@ -6,7 +6,6 @@ import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import edu.wpi.first.wpilibj2.command.StartEndCommand;
 import frc.robot.commands.pivot.PivotHomeCommand;
 import frc.robot.subsystems.ArmSubsystem;
-import frc.robot.constants.ArmConstants;
 import frc.robot.subsystems.IntakeSubsystem;
 import frc.robot.subsystems.PivotSubsystem;
 import frc.robot.subsystems.ShooterSubsystem;
@@ -19,7 +18,7 @@ public class HandoffCommandGroup extends SequentialCommandGroup {
 //                new StartEndCommand(() -> intakeSubsystem.roll(.60), intakeSubsystem::stop).withTimeout(.2),
                 new InstantCommand(() -> pivotSubsystem.setPivotPosition(0.5)).andThen(
                         new StartEndCommand(() -> armSubsystem.moveUp(.5), armSubsystem::stop).until(() ->
-                                armSubsystem.getEnc() <= ArmConstants.ARM_HANDOFF_POSITION).andThen(
+                                armSubsystem.getEnc() <= ArmSubsystem.ARM_HANDOFF_POSITION).andThen(
                                 new StartEndCommand(() -> shooterSubsystem.spinFeeder(-0.3),
                                         shooterSubsystem::stopFlywheel).alongWith(
                                         new StartEndCommand(() -> intakeSubsystem.roll(-.35), intakeSubsystem::stop))
