+++ src/main/java/frc/robot/commands/PivotLimitSwitchCommand.java
@@ -1,8 +1,6 @@
 package frc.robot.commands;
 
-import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.constants.PivotConstants;
 import frc.robot.subsystems.PivotSubsystem;
