--- src/main/java/frc/robot/constants/CANCoderConstants.java
@@ -1,6 +1,0 @@
-package frc.robot.constants;
-
-public final class CANCoderConstants {
-    public static final int COUNTS_PER_REV = 4096;
-    public static final double COUNTS_PER_DEG = COUNTS_PER_REV / 360.0;
-}
