--- src/main/java/frc/robot/constants/VisionConstants.java
@@ -1,13 +1,0 @@
-package frc.robot.constants;
-
-public final class VisionConstants {
-    public static final double LIMELIGHT_HEIGHT = 11.9; //placeholder
-    public static final double LIMELIGHT_X_TRANSLATION = 4.75; //placeholder
-    public static final double LIMELIGHT_Y_TRANSLATION = 0.0; //placeholder
-    public static final double LIMELIGHT_MOUNTING_ANGLE = 15.0; //placeholder
-
-    public static final String TABLE_NAME = "table";
-    public static final String LIMELIGHT_NAME = "limelight";
-
-    public static final String CAMERA_NAME = "camera";
-    }
