+++ src/main/java/frc/robot/subsystems/ArmSubsystem.java
@@ -1,58 +1,79 @@
 package frc.robot.subsystems;
 
 
-import com.ctre.phoenix6.configs.CANcoderConfiguration;
-import com.ctre.phoenix6.configs.TalonFXConfiguration;
-import com.ctre.phoenix6.controls.MotionMagicTorqueCurrentFOC;
-import com.ctre.phoenix6.controls.MotionMagicVelocityDutyCycle;
-import com.ctre.phoenix6.controls.MotionMagicVelocityVoltage;
+import com.ctre.phoenix6.configs.*;
 import com.ctre.phoenix6.controls.MotionMagicVoltage;
 import com.ctre.phoenix6.hardware.CANcoder;
 import com.ctre.phoenix6.hardware.TalonFX;
 import com.ctre.phoenix6.signals.*;
-import edu.wpi.first.wpilibj.DigitalInput;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.constants.ArmConstants;
-import frc.robot.constants.CANCoderConstants;
-import frc.robot.constants.TalonFXConstants;
-import org.opencv.core.Mat;
+import frc.robot.constants.ConfiguratorConstants;
 
 public class ArmSubsystem extends SubsystemBase {
 
     private final TalonFX armKraken;
     private final CANcoder armEncoder;
-    private ArmConstants.ArmPositions armPositions = ArmConstants.ArmPositions.STOWED;
+    public static final double GEAR_RATIO = 1.0 / (50.463 / 12.0);
+    public enum ArmPositions {
+        STOWED(90);
+        public final double angle;
+        public final double sensorUnits;
+        ArmPositions(double angle) {
+            this.angle = angle;
+            this.sensorUnits = angle / GEAR_RATIO * ConfiguratorConstants.COUNTS_PER_DEG;
+        }
+    }
+    private ArmPositions armPositions = ArmPositions.STOWED;
+    private static final double ARM_POSITION_STATUS_FRAME = 0.05;
+    private static final double ARM_VELOCITY_STATUS_FRAME = 0.01;
+    public static final double ARM_HANDOFF_POSITION = 0.03;
+    private static final double GRAVITY_FEEDFORWARD = 0.05;
+    private static final double ANGLE_TOLERANCE = 5;
+    public static final double ARM_P = 0;
+    public static final double ARM_I = 0;
+    public static final double ARM_D = 0;
+    public static final Slot0Configs SLOT_0_CONFIGS = new Slot0Configs().withKP(ARM_P).withKI(ARM_I).withKD(ARM_D).withGravityType(GravityTypeValue.Arm_Cosine);
+    public static final CurrentLimitsConfigs ARM_CURRENT_LIMIT = new CurrentLimitsConfigs().withSupplyCurrentLimitEnable(true).
+            withSupplyCurrentThreshold(55).withSupplyCurrentLimit(65).withSupplyTimeThreshold(0.1).withStatorCurrentLimitEnable(true).withStatorCurrentLimit(65);
+    public static final double MAGNET_OFFSET = -1; //placeholder
+    public static final InvertedValue ARM_INVERSION = InvertedValue.CounterClockwise_Positive;
+    public static final NeutralModeValue ARM_NEUTRAL_MODE = NeutralModeValue.Brake;
+    public static final SoftwareLimitSwitchConfigs ARM_SOFT_LIMIT = new SoftwareLimitSwitchConfigs().
+            withForwardSoftLimitEnable
+                    (true).
+            withForwardSoftLimitThreshold(
+                    0.0195 //placeholder
+            ).withReverseSoftLimitEnable(false).withReverseSoftLimitThreshold(
+                    65 //placeholder
+            );
 
     public ArmSubsystem() {
-        armKraken = new TalonFX(ArmConstants.ARM_KRAKEN_ID, TalonFXConstants.CANIVORE_NAME);
-        armEncoder = new CANcoder(ArmConstants.ARM_CANCODER_ID, TalonFXConstants.CANIVORE_NAME);
-
+        armKraken = new TalonFX(ConfiguratorConstants.ARM_KRAKEN_ID, ConfiguratorConstants.CANIVORE_NAME);
+        armEncoder = new CANcoder(ConfiguratorConstants.ARM_CANCODER_ID, ConfiguratorConstants.CANIVORE_NAME);
         var armConfigurator = armKraken.getConfigurator();
-        var talonFXConfiguration = new TalonFXConfiguration();
+        var armTalonFXConfiguration = new TalonFXConfiguration();
+        armTalonFXConfiguration.Feedback.FeedbackRemoteSensorID = armEncoder.getDeviceID();
+        armTalonFXConfiguration.Feedback.FeedbackSensorSource = FeedbackSensorSourceValue.RemoteCANcoder;
+        armTalonFXConfiguration.MotorOutput.Inverted = ARM_INVERSION;
+        armTalonFXConfiguration.MotorOutput.NeutralMode = ARM_NEUTRAL_MODE;
+        armTalonFXConfiguration.FutureProofConfigs = ConfiguratorConstants.TALON_FUTURE_PROOF;
+        armTalonFXConfiguration.Feedback.SensorToMechanismRatio = 50; //placeholder
+        armTalonFXConfiguration.Feedback.RotorToSensorRatio = 12.8;
+        armTalonFXConfiguration.CurrentLimits = ARM_CURRENT_LIMIT;
+        armTalonFXConfiguration.SoftwareLimitSwitch = ARM_SOFT_LIMIT;
+        armTalonFXConfiguration.Slot0 = SLOT_0_CONFIGS;
 
+        armKraken.getRotorVelocity().waitForUpdate(ARM_VELOCITY_STATUS_FRAME);
+        armKraken.getRotorPosition().waitForUpdate(ARM_POSITION_STATUS_FRAME);
 
-        talonFXConfiguration.Feedback.FeedbackRemoteSensorID = armEncoder.getDeviceID();
-        talonFXConfiguration.Feedback.FeedbackSensorSource = FeedbackSensorSourceValue.RemoteCANcoder;
-        talonFXConfiguration.MotorOutput.Inverted = ArmConstants.ARM_INVERSION;
-        talonFXConfiguration.MotorOutput.NeutralMode = ArmConstants.ARM_NEUTRAL_MODE;
-        talonFXConfiguration.FutureProofConfigs = TalonFXConstants.TALON_FUTURE_PROOF;
-        talonFXConfiguration.Feedback.SensorToMechanismRatio = 50; //placeholder
-        talonFXConfiguration.Feedback.RotorToSensorRatio = 12.8;
-        talonFXConfiguration.CurrentLimits = ArmConstants.ARM_CURRENT_LIMIT;
-        talonFXConfiguration.SoftwareLimitSwitch = ArmConstants.ARM_SOFT_LIMIT;
-        talonFXConfiguration.Slot0 = ArmConstants.SLOT_0_CONFIGS;
-
-        armKraken.getRotorVelocity().waitForUpdate(ArmConstants.ARM_VELOCITY_STATUS_FRAME);
-        armKraken.getRotorPosition().waitForUpdate(ArmConstants.ARM_POSITION_STATUS_FRAME);
-
-        armConfigurator.apply(talonFXConfiguration);
+        armConfigurator.apply(armTalonFXConfiguration);
 
         var armEncoderConfigurator = armEncoder.getConfigurator();
         var cancoderConfiguration = new CANcoderConfiguration();
 
         cancoderConfiguration.MagnetSensor.AbsoluteSensorRange = AbsoluteSensorRangeValue.Unsigned_0To1;
-        cancoderConfiguration.MagnetSensor.MagnetOffset = ArmConstants.MAGNET_OFFSET;
+        cancoderConfiguration.MagnetSensor.MagnetOffset = MAGNET_OFFSET;
         cancoderConfiguration.MagnetSensor.SensorDirection = SensorDirectionValue.Clockwise_Positive;
         armEncoderConfigurator.apply(cancoderConfiguration);
 
@@ -64,7 +85,7 @@ public class ArmSubsystem extends SubsystemBase {
     }
 
 
-    public void setPosition(ArmConstants.ArmPositions position) {
+    public void setPosition(ArmPositions position) {
         armPositions = position;
         setMotorsBrake();
 
@@ -72,7 +93,7 @@ public class ArmSubsystem extends SubsystemBase {
         double cosineScalar = Math.cos(radians);
 
         MotionMagicVoltage motionMagicVoltage = new MotionMagicVoltage(
-                position.sensorUnits, true, ArmConstants.GRAVITY_FEEDFORWARD * cosineScalar, 0,
+                position.sensorUnits, true, GRAVITY_FEEDFORWARD * cosineScalar, 0,
                 true, false, false);
 
         armKraken.setControl(motionMagicVoltage);
@@ -82,10 +103,10 @@ public class ArmSubsystem extends SubsystemBase {
     }
 
     public double getAngle() {
-        return armKraken.getRotorPosition().getValue() / CANCoderConstants.COUNTS_PER_DEG * ArmConstants.GEAR_RATIO;
+        return armKraken.getRotorPosition().getValue() / ConfiguratorConstants.COUNTS_PER_DEG * GEAR_RATIO;
     }
 
-    public ArmConstants.ArmPositions getArmPositions() {
+    public ArmPositions getArmPositions() {
         return armPositions;
     }
 
@@ -94,7 +115,7 @@ public class ArmSubsystem extends SubsystemBase {
     }
 
     public boolean isMotionFinished() {
-        return Math.abs(getAngle() - armPositions.angle) <= ArmConstants.ANGLE_TOLERANCE;
+        return Math.abs(getAngle() - armPositions.angle) <= ANGLE_TOLERANCE;
     }
 
     public void moveUp(double speed) {
