+++ src/main/java/frc/robot/subsystems/VisionSubsystem.java
@@ -3,7 +3,6 @@ package frc.robot.subsystems;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.constants.VisionConstants;
 import frc.robot.util.LimelightHelpers;
 
 // does this need to be a subsystem? we aren't using any commands
@@ -11,29 +10,31 @@ public class VisionSubsystem extends SubsystemBase {
     public double xFinal;
     public double yFinal;
 
+    public static final String LIMELIGHT_NAME = "limelight";
+
 
     public boolean hasTargets() {
-        return LimelightHelpers.getTV(VisionConstants.LIMELIGHT_NAME);
+        return LimelightHelpers.getTV(LIMELIGHT_NAME);
     }
 
     public double getX() {
-        return LimelightHelpers.getTX(VisionConstants.LIMELIGHT_NAME);
+        return LimelightHelpers.getTX(LIMELIGHT_NAME);
     }
 
     public double getY() {
-        return LimelightHelpers.getTY(VisionConstants.LIMELIGHT_NAME);
+        return LimelightHelpers.getTY(LIMELIGHT_NAME);
     }
 
     public double[] getPose() {
-        return LimelightHelpers.getBotPose(VisionConstants.LIMELIGHT_NAME);
+        return LimelightHelpers.getBotPose(LIMELIGHT_NAME);
     }
 
     public Pose3d getPose3d() {
-        return LimelightHelpers.getBotPose3d(VisionConstants.LIMELIGHT_NAME);
+        return LimelightHelpers.getBotPose3d(LIMELIGHT_NAME);
     }
 
     public Pose2d getPose2d() {
-        return LimelightHelpers.getBotPose2d(VisionConstants.LIMELIGHT_NAME);
+        return LimelightHelpers.getBotPose2d(LIMELIGHT_NAME);
     }
 
 
@@ -46,22 +47,22 @@ public class VisionSubsystem extends SubsystemBase {
     }
 
     public double getID() {
-        return LimelightHelpers.getFiducialID(VisionConstants.LIMELIGHT_NAME);
+        return LimelightHelpers.getFiducialID(LIMELIGHT_NAME);
     }
 
     public void setPipeline(int num) {
-        LimelightHelpers.setPipelineIndex(VisionConstants.LIMELIGHT_NAME, num);
+        LimelightHelpers.setPipelineIndex(LIMELIGHT_NAME, num);
     }
 
     public LimelightHelpers.Results getTargetingResults() {
-        return LimelightHelpers.getLatestResults(VisionConstants.LIMELIGHT_NAME).targetingResults;
+        return LimelightHelpers.getLatestResults(LIMELIGHT_NAME).targetingResults;
     }
 
     public void ledOff() {
-        LimelightHelpers.setLEDMode_ForceOff(VisionConstants.LIMELIGHT_NAME);
+        LimelightHelpers.setLEDMode_ForceOff(LIMELIGHT_NAME);
     }
 
     public void ledOn() {
-        LimelightHelpers.setLEDMode_ForceOn(VisionConstants.LIMELIGHT_NAME);
+        LimelightHelpers.setLEDMode_ForceOn(LIMELIGHT_NAME);
     }
 }
